refactor(#792): extract dsp_worker unit tests into dsp_worker_tests.rs (sweep)
dsp_worker.rs 719 -> 542 (leaves the debt list). Moved the two inline test
modules (budget_tracker_tests, saturation_recovery_tests) to a sibling file
wired via #[path]; their super:: references become super::super:: (descendant
modules still see the parent's private items). infra-cpal tests pass, zero warnings.
